+++ examples/Application/String/MainApplication.cpp
@@ -4,7 +4,7 @@ class MainApplication {
 public:
 		static void main(Array<String> arguments) {
 			System::out::println("N = 10^8");
-			register int i;
+			register long int i;
 			for (i=0; i<1000000000; i++) {
 				String str;
 				str += "hello world";
@@ -12,6 +12,6 @@ public:
 		}
 };
 
-int main(int argc, char **argv) {
+long int main(long int argc, char **argv) {
 	return Application(MainApplication::main, argc, argv);
 }
